Refatorar geração de figuras em módulos reutilizáveis
